[bindings] Add python bindings for ControlPolyZero and ControlPolyOne

+++ include/crocoddyl/core/controls/poly-one.hxx
@@ -79,7 +79,7 @@ void ControlPolyOneTpl<Scalar>::multiplyByDValue(double t, const Eigen::Ref<cons
     throw_pretty("Invalid argument: "
                 << "p has wrong dimension (it should be " + std::to_string(np_) + ")");
   }
-  if (A.rows() != out.rows() || A.cols()!=nu_ || out.cols()!=np_) {
+  if (A.rows() != out.rows() || static_cast<std::size_t>(A.cols())!=nu_ || static_cast<std::size_t>(out.cols())!=np_) {
     throw_pretty("Invalid argument: "
                 << "A and out have wrong dimensions (" + std::to_string(A.rows()) + "," + std::to_string(A.cols()) 
                 + " and " + std::to_string(out.rows()) + "," + std::to_string(out.cols()) + + ")");
